add gain(const)

+++ L3DG20.hpp
@@ -13,7 +13,7 @@ class L3DG20 : private RPi_i2c
   int read_xyz(const int reg);
 public:
   L3DG20(void);
-  int read_x(void);
-  int read_y(void);
-  int read_z(void);
+  double read_x(void);
+  double read_y(void);
+  double read_z(void);
 };
